+++ source/compiler/sc2.c
@@ -377,7 +377,7 @@ static void readline(unsigned char *line)
           *ptr='\0';    /* erase '\n' (and any trailing whitespace) */
         } /* if */
       } /* if */
-      num-=strlen((char*)line);
+      num-=(int)strlen((char*)line);
       line+=strlen((char*)line);
     } /* if */
     fline+=1;
@@ -1385,7 +1385,7 @@ static int command(void)
   case tpUNDEF:
     if (!SKIPPING) {
       if (lex(&val,&str)==tSYMBOL) {
-        ret=delete_subst(str,strlen(str));
+        ret=delete_subst(str,(int)strlen(str));
         if (!ret) {
           /* also undefine normal constants */
           symbol *sym=findconst(str,NULL);
@@ -1632,7 +1632,7 @@ static int substpattern(unsigned char *line,size_t buffersize,char *pattern,char
         arg=*(e+1)-'0';
         assert(arg>=0 && arg<=9);
         if (args[arg]!=NULL)
-          len+=strlen((char*)args[arg]);
+          len+=(int)strlen((char*)args[arg]);
         else
           len+=2;     /* copy '%' plus digit */
         e++;          /* skip %, digit is skipped later */
@@ -2699,7 +2699,7 @@ SC_FUNC void delete_symbols(symbol *root,int level,int delete_labels,int delete_
 SC_FUNC uint32_t namehash(const char *name)
 {
   const unsigned char *ptr=(const unsigned char *)name;
-  int len=strlen(name);
+  int len=(int)strlen(name);
   if (len==0)
     return 0L;
   assert(len<256);
